adk_agent/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-adk
google-cloud-bigquery